// src/types/supabaseTypes.ts
export interface definitions {
  comments: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    id: string;
    /**
     * Note:
     * This is a Foreign Key to `profiles.user_id`.<fk table='profiles' column='user_id'/>
     */
    user_id: string;
    created_at: string;
    updated_at: string;
    /**
     * Note:
     * This is a Foreign Key to `posts.id`.<fk table='posts' column='id'/>
     */
    post_id: string;
    text: string;
  };
  favorits: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    id: string;
    /**
     * Note:
     * This is a Foreign Key to `profiles.user_id`.<fk table='profiles' column='user_id'/>
     */
    user_id?: string;
    created_at: string;
    updated_at: string;
    /**
     * Note:
     * This is a Foreign Key to `posts.id`.<fk table='posts' column='id'/>
     */
    post_id: string;
  };
  likes: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    id: string;
    /**
     * Note:
     * This is a Foreign Key to `profiles.user_id`.<fk table='profiles' column='user_id'/>
     */
    user_id?: string;
    created_at: string;
    updated_at: string;
    /**
     * Note:
     * This is a Foreign Key to `posts.id`.<fk table='posts' column='id'/>
     */
    post_id: string;
  };
  posts: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    id: string;
    /**
     * Note:
     * This is a Foreign Key to `profiles.user_id`.<fk table='profiles' column='user_id'/>
     */
    user_id?: string;
    created_at: string;
    updated_at: string;
    text: string;
  };
  profiles: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    user_id: string;
    created_at: string;
    updated_at: string;
    name?: string;
    avatar?: string;
    wedding_hall?: string;
    description?: string;
  };
  todos: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    id: number;
    /**
     * Note:
     * This is a Foreign Key to `profiles.user_id`.<fk table='profiles' column='user_id'/>
     */
    user_id?: string;
    created_at: string;
    updated_at: string;
    group_id?: number;
    todo: string;
    is_done?: boolean;
  };
}